# docs/swagger.yaml
basePath: /
definitions:
...
  structs.Credentials:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  structs.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
    type: object
...
  structs.User:
    properties:
      created_at: {}
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at: {}
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Casheex is a simple RESTful API for a casheer system, designed to support
    basic cashier operations in retail or small business environments.
  title: Casheex
  version: "1.0"
paths:
...
  /api/users:
    get:
      description: Retrieve a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Store a new user into the database
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/structs.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - Users
  /api/users/{id}:
    delete:
      description: Delete a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      description: Retrieve a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Find user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user data by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/structs.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - Users
  /api/users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: User credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/structs.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
      summary: User login
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"